backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.35.0
torch==2.1.0
waitress==2.1.2
requests==2.31.0
pytest==7.4.3
black==23.11.0